+++ project/jobs/edfi_api_to_amt.py
@@ -159,8 +159,8 @@ edfi_api_dev_job = edfi_api_to_amt.to_job(
             "base_url": os.getenv("EDFI_BASE_URL"),
             "api_key": os.getenv("EDFI_API_KEY"),
             "api_secret": os.getenv("EDFI_API_SECRET"),
-            "api_page_limit": 5000,
-            "api_mode": "YearSpecific" # DistrictSpecific, SharedInstance, YearSpecific
+            "api_page_limit": 100,
+            "api_mode": "SharedInstance" # DistrictSpecific, SharedInstance, YearSpecific
         }),
         "warehouse": bq_client.configured({
             "staging_gcs_bucket": os.getenv("GCS_BUCKET_DEV"),
@@ -176,7 +176,7 @@ edfi_api_dev_job = edfi_api_to_amt.to_job(
         "inputs": {
             "edfi_api_endpoints": { "value": edfi_api_endpoints },
             "school_year": { "value": 2022 },
-            "use_change_queries": { "value": True }
+            "use_change_queries": { "value": False }
         }
     },
 )