+++ project_dbt/models/marts/attendance/rpt_student_attendance.sql
@@ -1,4 +1,14 @@
 
+WITH max_school_year_dates AS (
+
+    SELECT
+        school_year,
+        MAX(date) AS latest_date
+    FROM {{ ref('fct_student_attendance') }}
+    GROUP BY 1
+
+)
+
 SELECT
     fct_student_attendance.school_year                                     AS school_year,
     dim_student.student_unique_id                                          AS student_unique_id,
@@ -6,7 +16,7 @@ SELECT
     dim_school.school_key                                                  AS school_key,
     dim_school.school_id                                                   AS school_id,
     dim_date.date                                                          AS date,
-    dim_date.month                                                         AS month,
+    dim_date.month_name                                                    AS month_name,
     dim_date.month_sort_order                                              AS month_sort_order,
     fct_student_attendance.school_attendance_event_category_descriptor     AS school_attendance_event_category_descriptor,
     fct_student_attendance.event_duration                                  AS event_duration,
@@ -14,14 +24,18 @@ SELECT
     fct_student_attendance.reported_as_absent_from_school                  AS reported_as_absent_from_school,
     fct_student_attendance.reported_as_present_at_home_room                AS reported_as_present_at_home_room,
     fct_student_attendance.reported_as_absent_from_home_room               AS reported_as_absent_from_home_room,
-    fct_student_attendance.reported_as_is_present_in_all_sections          AS reported_as_is_present_in_all_sections,
-    fct_student_attendance.reported_as_absent_from_any_section             AS reported_as_absent_from_any_section,
+    fct_student_attendance.is_on_the_verge                                 AS is_on_the_verge,
+    fct_student_attendance.is_chronically_absent                           AS is_chronically_absent,
+    IF(
+        dim_date.date = max_school_year_dates.latest_date, TRUE, FALSE
+    )                                                                      AS is_latest_date_avaliable,
     dim_local_education_agency.local_education_agency_name                 AS local_education_agency_name,
     dim_school.school_name                                                 AS school_name,
     dim_student.student_last_surname                                       AS student_last_surname,
     dim_student.student_first_name                                         AS student_first_name,
     dim_student.is_actively_enrolled                                       AS is_actively_enrolled,
     dim_student.grade_level                                                AS grade_level,
+    dim_student.grade_level_id                                             AS grade_level_id,
     dim_student.gender                                                     AS gender,
     dim_student.limited_english_proficiency                                AS limited_english_proficiency,
     dim_student.is_english_language_learner                                AS is_english_language_learner,
@@ -37,3 +51,4 @@ LEFT JOIN {{ ref('dim_school') }} dim_school
     ON fct_student_attendance.school_key = dim_school.school_key
 LEFT JOIN {{ ref('dim_local_education_agency') }} dim_local_education_agency
     ON dim_school.local_education_agency_key = dim_local_education_agency.local_education_agency_key
+LEFT JOIN max_school_year_dates ON fct_student_attendance.school_year = max_school_year_dates.school_year